Restructuring the main coordinator to ensure that unregistered users exist before attempting to enter the conversation list.
