+++ iOS/API/GetChatToken+Cloud.swift
@@ -7,8 +7,6 @@
 //
 
 import Foundation
-import Parse
-import Combine
 
 struct GetChatToken: CloudFunction {
